refactor: moved the passkey section to a new file
